feat: Add tests for abandoned cart notification and webhook handling
