+++ retail/vtex/tests/webhooks/test_abandoned_cart_notification_view.py
