+++ retail/vtex/tests/webhooks/test_serializers.py
